cargo_api_project/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.0.6
django-froala-editor==4.0.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.12.0
idna==3.4
joblib==1.2.0
numpy==1.24.3
Pillow==9.2.0
platformdirs==3.5.1
# psycopg2==2.9.6
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.14.0
threadpoolctl==3.1.0
django-location-field== 2.7.1
urllib3==1.25.11
virtualenv==20.23.0 
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
sqlparse==0.4.2
tzdata==2021.5
mysqlclient==2.1.1
googlemaps ==4.10.0
django-google-maps ==0.13.0
firebase-admin